Add reuse to dep not-a-resolver
This has two pieces that work in conjunction:

1. Whenever picking versions, if we've already chosen one during this
   call to `walk()` then prioritize keeping that.
2. If we need to pick a version greenfield, prioritize keeping the
   version we currently have, if any, otherwise choose the newest.

You can try this out in the CLI by passing e.g. `--have numpy==1.19.0`.

If no. 1 above is impossible, it will output a message and keep going.
Only a couple of real-world examples I've seen trigger this behavior,
including https://gist.github.com/thatch/9c0e076f96569067851fbf8c43d39c53
(but not fixing the edge that shows line 5, after being revised for line
20, is a bug).

If you pass a non-public version which can be used in no. 2, it will be
displayed as no-sdist no-wheel which is incorrect (it probably deserves
a third flag for "reused" or so, which would be useful to set even for
public versions).  This section of the code needs a rewrite.

Versions reused from no. 2 will not have deps listed, initially because
non-public versions don't have easily-accessible deps, but also for
consistency and probably not needing them (if this is being used to
determine what new needs to be installed).