Make revs sync
There were some unclosed socket warnings on 3.12 that I can't
immediately track down.  I'm slowly moving away from using async for
this project as it is troublesome to parallelize in the `deps`
subcommand, troublesome to test, and doesn't add a lot of value for
commands that just check one project.